requirements.txt
cassandra-driver==3.29.2
click==8.1.8
geomet==0.2.1.post1
python-dotenv==1.0.1
pyzmq==26.2.0
six==1.17.0
